==> Scripts/1 - Data Wrangling.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

path = r'C:\Users\Ben\Desktop\Career Foundry Tasks\Instacart Basket Analysis 02-2023'

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

df_ords.drop(columns = ['eval_set'])

df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

df_dep.T

df_dep_t = df_dep.T

df_dep_t.reset_index()

new_header = df_dep_t.iloc[0]

df_dep_t_new = df_dep_t[1:]

df_dep_t_new.columns = new_header

data_dict = df_dep_t_new.to_dict('index')

print(data_dict.get('19'))

df_snacks =  df_prods[df_prods['department_id']==19]

df_prods['department_id']==19

df_prods[df_prods['department_id']==19]

df_snacks.head()

# Task

df_ords

df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords.rename(columns = {'days_since_prior_order' : 'days_since_last_order'}, inplace = True)

df_ords['order_hour_of_day'].value_counts(dropna = False)

10 is the most frequent hour at 288418 orders

print(data_dict.get('4'))

data_dict

df_breakfast =  df_prods[df_prods['department_id']==14]

df_breakfast

df_dinner_parties = df_prods.loc[df_prods['department_id'].isin([5,20,7,12])]

df_dinner_parties

7650 rows from the last dataframe

df_ords_user_1 = df_ords.loc[df_ords['user_id'].isin([1])]

df_ords

df_ords_user_1 = df_ords.loc[df_ords['user_id'].isin([1])]

df_ords_user_1

df_ords_user_1.describe()

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))
