==> Scripts/2 - Data Consistency Checks.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

path = r'C:\Users\Ben\Desktop\Career Foundry Tasks\Instacart Basket Analysis 02-2023'

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

# Create a dataframe
df_test = pd.DataFrame()

# Create a mixed type column
df_test['mix'] = ['a', 'b', 1, True]

df_test.head()

# Check for mixed types
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

df_test['mix'] = df_test['mix'].astype('str')

df_prods.isnull().sum()

df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

df_prods.shape

df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'))

# Task

df_ords.describe()

Max values are really low. The hour of the day max is only 2.3, but there are 24 hours in a day. Days since prior order is also low at 3, especially when the 50% is 7.

# Check for mixed data
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

No columns shown so there are no columns with mixed-type data

# Check for missing data
df_ords.isnull().sum()

The only column with missing values is days since prior order. This is likely just first time customers who haven't ordered anything before so the missing values aren't true missing data points.

You could address the missing values by removing the null values by doing the following: df_ords_clean = df_ords[df_ords['days_since_prior_order'].isnull() == False]
However, the missing values are data points showing that a customer is new so I wouldn't actually drop them.

# Check for duplicates
df_dups_ords = df_ords[df_ords.duplicated()]

df_dups_ords

No duplicates were found.

df_dups_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_checked.csv'))
