==> Scripts/3 - Data Merge.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

path = r'C:\Users\Ben\Desktop\Career Foundry Tasks\Instacart Basket Analysis 02-2023'

# Importing merged dataframe
df_ords_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

# Check shape
df_ords_merge.shape

# Same shape as the exported data

# Importing prepared products dataframe
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index_col = False)

df_ords_merge.head()

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order', 'product_id', 'add_to_cart_order', 'reordered']

# Filtering out eval_set and _merge columns to merge new dataframe
df_ords_merge2 = df_ords_merge.filter(vars_list)

df_ords_merge2.head()

df_ords_merge2.shape

df_prods.head()

df_prods.shape

# Merge dataframes
df_ords_prods = df_ords_merge2.merge(df_prods, on = 'product_id', indicator = True)

# Results of merge using the merge flag
df_ords_prods['_merge'].value_counts()

# Export new merged dataframe
df_ords_prods.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_full_combined.pkl'))
