==> Scripts/4 - Deriving New Variables.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

path = r'C:\Users\Ben\Desktop\Career Foundry Tasks\Instacart Basket Analysis 02-2023'

df_ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_full_combined.pkl'))

df = df_ords_prods_merge[:1000000]

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

df['price_range'] = df.apply(price_label, axis=1)

df['price_range'].value_counts(dropna = False)

df['prices'].max()

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna = False)

df_ords_prods_merge.loc[df_ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_ords_prods_merge.loc[(df_ords_prods_merge['prices'] <= 15) & (df_ords_prods_merge['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df_ords_prods_merge.loc[df_ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prods_merge['price_range_loc'].value_counts(dropna = False)

df_ords_prods_merge['order_dow'].value_counts(dropna = False)

result = []

for value in df_ords_prods_merge["order_dow"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

df_ords_prods_merge['busiest_day'] = result

df_ords_prods_merge['busiest_day'].value_counts(dropna = False)

# Task

# Busiest days

result = []

for value in df_ords_prods_merge["order_dow"]:
  if value == 0:
    result.append("Busiest days")
  elif value == 1:
    result.append("Busiest days")    
  elif value == 4:
    result.append("Slowest days")
  elif value == 3:
    result.append("Slowest days")    
  else:
    result.append("Regularly busy")

# Add new column to dataframe

df_ords_prods_merge['busiest_days'] = result

# Print new column

result

#Frequency of new column

df_ords_prods_merge['busiest_days'].value_counts(dropna = False)

Busiest days should be a sum of the highest two days and the slowest days is a sum of the lowest two days. When checking the frequency for each day of the week, these numbers are the sum of the highest/lowest 2.

df_ords_prods_merge.head()

# Frequency of order hour column

df_ords_prods_merge['order_hour_of_day'].value_counts(dropna = False)

# New column for busiest hours of the day by splitting in three groups of 8 hours each

result = []

for value in df_ords_prods_merge["order_hour_of_day"]:
  if value in [10, 11, 14, 15, 13, 12, 16, 9]:
    result.append("Most orders")
  elif value in [3, 4, 2, 5, 1, 0, 6, 23]:
    result.append("Fewest orders")    
  else:
    result.append("Average orders")

# Print new column

result

# Add new column to dataframe

df_ords_prods_merge['busiest_hours_of_the_day'] = result

# Frequency of new column

df_ords_prods_merge['busiest_hours_of_the_day'].value_counts(dropna = False)

# Export dataframe

df_ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_new_variables.pkl'))
