==> Scripts/5 - Grouping Data.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

path = r'C:\Users\Ben\Desktop\Career Foundry Tasks\Instacart Basket Analysis 02-2023'

ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_new_variables.pkl'))

ords_prods_merge.groupby('department_id').agg({'order_number': ['mean']})

Because this is the full dataset and not a subset, all department ids are listed. The mean of the same department ids is different here than the subset because it is including all data points. The subset mean compared to the full dataset mean is quite different. More of a difference than I would have expected, for example department 14 mean is 19.2 in the subset but only 16.7 with the full data.

ords_prods_merge.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

ords_prods_merge.head()

#Spending habits of the three types of customers
ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean', 'min', 'max']})

#Average price spending flag
ords_prods_merge['average_price'] = ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

ords_prods_merge.loc[ords_prods_merge['average_price'] < 10, 'spender_flag'] = 'Low spender'

ords_prods_merge.loc[ords_prods_merge['average_price'] >= 10, 'spender_flag'] = 'High spender'

ords_prods_merge['spender_flag'].value_counts(dropna = False)

#Frequency flag
ords_prods_merge['median_prior_order'] = ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

ords_prods_merge.loc[ords_prods_merge['median_prior_order'] > 20, 'frequency_flag'] = 'Non-frequent customer'

ords_prods_merge.loc[(ords_prods_merge['median_prior_order'] <= 20) & (ords_prods_merge['median_prior_order'] < 10), 'frequency_flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['median_prior_order'] <= 10, 'frequency_flag'] = 'Frequent customer'

ords_prods_merge['frequency_flag'].value_counts(dropna = False)

#Export dataframe
ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_agg.pkl'))
