==> Scripts/6 - Data Visualization Part 1.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

path = r'C:\Users\Ben\Desktop\Career Foundry Tasks\Instacart Basket Analysis 02-2023'

# Customer dataframe

ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_new_variables.pkl'))

df_customer.head()

df_customer.rename(columns = {'Surnam' : 'Last Name'}, inplace = True)

df_customer.rename(columns = {'STATE' : 'State'}, inplace = True)

df_customer.rename(columns = {'fam_status' : 'marital_status'}, inplace = True)

df_customer.head()

df_customer.describe()

# Check for missing values

df_customer.isnull().sum()

df_nan = df_customer[df_customer['First Name'].isnull() == True]

# Check the rows with null values

df_nan

All null values are rows where the first name is missing. There is enough data on those rows where the data can still be included.

# Check for duplicates

df_dups = df_customer[df_customer.duplicated()]

df_dups

No duplicates found

ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_agg.pkl'))

ords_prods_merge.head()

# Merging customer data with instacart data

df_ords_prods_cust = ords_prods_merge.merge(df_customer, on = 'user_id', indicator = True)

ords_prods_merge.rename(columns = {'_merge' : 'merge'}, inplace = True)

df_ords_prods_cust.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_cust.pkl'))
