==> Scripts/8 - Final Analysis Part 1.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

path = r'C:\Users\Ben\Desktop\Career Foundry Tasks\Instacart Basket Analysis 02-2023'

ords_prods_cust = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_cust.pkl'))

# Security Implications

PII data exists because the data includes the names of people.

ords_prods_cust.head()

ords_prods_cust['State'].value_counts().sort_index()

# Assign States to Regions

NE = ["Maine", "New Hampshire", "Vermont", "Massachusetts", "Rhode Island", "Connecticut", "New York", "Pennsylvania", "New Jersey"]
MW = ["Wisconsin", "Michigan", "Illinois", "Indiana", "Ohio", "North Dakota", "South Dakota", "Nebraska", "Kansas", "Minnesota", "Iowa", "Missouri"]
S = ["Delaware", "Maryland", "District of Columbia", "Virginia", "West Virginia", "North Carolina", "South Carolina", "Georgia", "Florida", "Kentucky", "Tennessee", "Mississippi", "Alabama", "Oklahoma", "Texas", "Arkansas", "Louisiana"]
W = ["Idaho", "Montana", "Wyoming", "Nevada", "Utah", "Colorado", "Arizona", "New Mexico", "Alaska", "Washington", "Oregon", "California", "Hawaii"]

# Create Region column with the regions assigned based off of the state

ords_prods_cust.loc[ords_prods_cust['State'].isin(NE), 'Region'] = 'Northeast'
ords_prods_cust.loc[ords_prods_cust['State'].isin(MW), 'Region'] = 'Midwest'
ords_prods_cust.loc[ords_prods_cust['State'].isin(S), 'Region'] = 'South'
ords_prods_cust.loc[ords_prods_cust['State'].isin(W), 'Region'] = 'West'

# Verify all data has a valid region

ords_prods_cust['Region'].value_counts(dropna = False)

# Spending habits based off of region

crosstab = pd.crosstab(ords_prods_cust['Region'], ords_prods_cust['spender_flag'], dropna = False)

crosstab

ords_prods_cust.info()

ords_prods_cust.loc[ords_prods_cust['max_order'] < 5, 'exclusion_flag'] = 'exclude'

ords_prods_cust.loc[ords_prods_cust['max_order'] >= 5, 'exclusion_flag'] = 'include'

ords_prods_cust['exclusion_flag'].value_counts(dropna = False)

df_high_activity = ords_prods_cust.loc[ords_prods_cust['exclusion_flag'] == 'include']

df_high_activity.info()

df_low_activity = ords_prods_cust.loc[ords_prods_cust['exclusion_flag'] == 'exclude']

df_high_activity.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'high_activity.pkl'))

df_low_activity.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'low_activity.pkl'))

df_high_activity[["Age", "income", "n_dependants"]].describe()

ords_prods_cust['order_dow'].value_counts(dropna = False)
